// src/store/types/Video/VideoFormat.ts
// Generated by https://quicktype.io

export interface VideoFormat {
    height?:             string;
    tbr?:                number;
    ext:                string;
    http_headers?:       HTTPHeaders;
    asr?:                number;
    filesize?:           number;
    acodec?:             string;
    fps?:                string;
    protocol?:           string;
    player_url?:         null;
    downloader_options?: DownloaderOptions;
    width?:              string;
    vcodec?:             string;
    abr?:                number;
    format:             string;
    format_note:        string;
    format_id:          string;
    url?:                string;
    quality?:            number | undefined
}

export interface DownloaderOptions {
    http_chunk_size: number;
}

export interface HTTPHeaders {
    'User-Agent':      string;
    'Accept-Encoding': string;
    'Accept-Language': string;
    'Accept-Charset':  string;
    Accept:            string;
}